restapi/requirements.txt
greenlet
PasteDeploy
eventlet
Routes
WebOb
